+++ src/Game/Resources/CombinedResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BinaryStudioAcademy\Game\Resource;
+namespace BinaryStudioAcademy\Game\Resources;
 
 abstract class CombinedResource extends Resource
 {
